+++ api/v1/helpers_test.go
@@ -13,6 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+
 package v1
 
 import (
@@ -27,7 +28,7 @@ type test struct {
 	backend               Backend
 	expectedContainer     string
 	expectedPrefix        string
-	expectedMaxConnection int
+	expectedMaxConnection int64
 	expectedLocation      string
 	expectedProvider      string
 	expectedEndpoint      string
